Use an atomic.Pointer to store wrapped DB objects
This allows atomically swapping to the new object. A timeout is
introduced so that if the object is swapped after the existing one has
been looked up it isn't immediately invalidated. The timeout is
arbitrary but has been set to five minutes as anything running that long
probably _should_ be interrupted.

This could be done with mutexes to avoid the arbitrary timeout, but
that carries its own risk: if a write lock is attempted to be acquired
while a query is holding a read lock, the write lock will block, and any
subsequent read locks will also block since write locks take priority.
As a result, if a query is long-running or timing out, all other queries
in the system will become blocked. Our default user request timeout is
90 seconds and some jobs have no timeout at all, so this is not a purely
theoretical concern, even if it's unlikely. So for the moment, using an
atomic pointer feels gross but safer.