refact(cmd): Add SetupWorkerInitialUpstreams to config.Config
This moves the logic that determines the worker initial upstreams in the
case where the server is running as both a controller and a worker into
a method on the config.Config. This will allow this step to be used more
easily in more places, notably in the SIGHUP reload behavior.