refact(genapi): Change sliceSubtypes into subactions
sliceSubtypes is currently used to generate api client code to call some
custom actions for a given resource within Boundary's API.
(eg: Target's SetCredentialSources)

This commit changes sliceSubtypes to subactions to make this genapi
option a little more broad, as it can already be used for non-slice
"subtypes". Actually, it doesn't even need to be a "type" at all.