feat(genapi): Custom operations per subaction
Support specifying which operations to create for any given subaction.
This will be required for upcoming work as a Target address does not
support the "Add" operation.

Sidenote: The *.gen.go file changes in this commit are due to the
operation funcs being moved around because of the change in the
template. Before, we were ranging through the operations, then
subactions. Now, we're ranging through the subactions, then ops. This
changes the order in which the funcs appear in the files from:

(in order)
AddX, AddY, SetX, SetY, RemoveX, RemoveY ...

to

(in order)
AddX, SetX, RemoveX, AddY, SetY, RemoveY ...

But the functions present and their content is the same.