fix(base): Merge all config wrappers of same purpose
Previously, multiple wrappers of the same purpose would
result in the last wrapper of each purpose being used. This
meant it was impossible to transition from one root wrapper
to another. We not merge all specified wrappers, with the
later wrappers becoming the wrapper used for encryption.
This allows users to migrate from one kms to another gradually.