Fix string regression in attribute parsing
In #2721, the ability for key-only values was added via using a
wrapperspb.StringValue that could be null. This introduced a bug because
originally the default case for `-attr` was to get the string value but
instead it used the incoming value as-is to allow for nulls; the null
check was moved up later but the original behavior was not restored. As
a result, non-explicitly-quoted strings carried through the
wrapperspb.StringValue, resulting in a map for the value in JSON instead
of a string.

This also removes the ability to set explicitly quoted strings via the
CLI. We had a mismatch in behavior where when it was caught one way it
would trim those quotes; another way it would not, so it standardizes on
removing leading/trailing quote marks.