chore(enos): Add debug variable
Generally, if there's failure in the test suite for any reason, enos/terraform will throw an error and you
would not be able to access the environment variables needed to test locally. Enabling this
variable will ensure that the enos scenario passes so that you can run `test_e2e_env.sh` to extract the
environment variables.