fix(perms): Take ids of applicable grants into account when permission building
An "applicable" grant is a grant that matches the requested resource, or "*",
and has one or more actions that exist in the list of a resource's individual
actions (the actions that can be performed on a single instance of a resource,
IdActions), or "*".

This also fixes the tests which had action.List on their action set
for almost all test cases. This isn't the case because list isn't
an action that can be performed on an invidual resource (IdActions).